add Base class for generic hostile AI
All new hostile behaviors are meant to inherit from this one
