+++ UnityProject/Assets/Resources/Prefabs/Mobs/NPC/Resources/Hostile/Carp/NPC_Carp_Regular.prefab
@@ -45,6 +45,7 @@ SpriteRenderer:
   m_MotionVectors: 1
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
   m_RenderingLayerMask: 4294967295
   m_RendererPriority: 0
   m_Materials:
@@ -135,6 +136,17 @@ MonoBehaviour:
   isBleeding: 0
   livingHealthBehaviour: {fileID: 0}
   Severity: 0
+  armor:
+    Melee: 0
+    Bullet: 0
+    Laser: 0
+    Energy: 0
+    Bomb: 0
+    Rad: 0
+    Fire: 0
+    Acid: 0
+    Magic: 0
+    Bio: 0
 --- !u!1 &4826259315037707297
 GameObject:
   m_ObjectHideFlags: 0
@@ -190,6 +202,17 @@ MonoBehaviour:
   isBleeding: 0
   livingHealthBehaviour: {fileID: 0}
   Severity: 0
+  armor:
+    Melee: 0
+    Bullet: 0
+    Laser: 0
+    Energy: 0
+    Bomb: 0
+    Rad: 0
+    Fire: 0
+    Acid: 0
+    Magic: 0
+    Bio: 0
 --- !u!1 &4827224533573323059
 GameObject:
   m_ObjectHideFlags: 0
@@ -246,9 +269,9 @@ GameObject:
   - component: {fileID: 3493423571535560689}
   - component: {fileID: 3651291330810827203}
   - component: {fileID: 3708589072684735304}
-  - component: {fileID: 6990826745820629644}
   - component: {fileID: 4145711099332912043}
   - component: {fileID: 5083886736440002245}
+  - component: {fileID: -4295618840045907110}
   m_Layer: 12
   m_Name: NPC_Carp_Regular
   m_TagString: Untagged
@@ -454,6 +477,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  isDirty: 0
   sceneId: 0
   serverOnly: 0
   m_AssetId: 
@@ -563,26 +587,6 @@ MonoBehaviour:
   attackVerb: slashed
   defaultTarget: 0
   meleeCoolDown: 0.4
---- !u!114 &6990826745820629644
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4827737755655954553}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b9b02e30a4b623144905790b2eaf49b9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  mobName: Carp
-  knockedDownRotation: 90
-  searchTickRate: 0.5
-  DeathSounds: []
-  GenericSounds:
-  - bite
-  PlaySoundTime: 3
-  status: 0
 --- !u!114 &4145711099332912043
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -651,3 +655,23 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   AnimatorElement: 1
   LoopEvery: 2
+--- !u!114 &-4295618840045907110
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4827737755655954553}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 243a3c078e195af4dadbc9f673627997, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  mobName: Space carp
+  knockedDownRotation: 90
+  deathSounds: []
+  randomSound: []
+  playRandomSoundTimer: 3
+  randomSoundProbability: 20
+  searchTickRate: 0.5
+  currentStatus: 0
