+++ UnityProject/Assets/Resources/Prefabs/Mobs/NPC/Resources/Hostile/Xenomorph/NPC_Xenomorph.prefab
@@ -45,6 +45,7 @@ SpriteRenderer:
   m_MotionVectors: 1
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
   m_RenderingLayerMask: 4294967295
   m_RendererPriority: 0
   m_Materials:
@@ -135,6 +136,17 @@ MonoBehaviour:
   isBleeding: 0
   livingHealthBehaviour: {fileID: 0}
   Severity: 0
+  armor:
+    Melee: 0
+    Bullet: 0
+    Laser: 0
+    Energy: 0
+    Bomb: 0
+    Rad: 0
+    Fire: 0
+    Acid: 0
+    Magic: 0
+    Bio: 0
 --- !u!1 &4826259315037707297
 GameObject:
   m_ObjectHideFlags: 0
@@ -190,6 +202,17 @@ MonoBehaviour:
   isBleeding: 0
   livingHealthBehaviour: {fileID: 0}
   Severity: 0
+  armor:
+    Melee: 0
+    Bullet: 0
+    Laser: 0
+    Energy: 0
+    Bomb: 0
+    Rad: 0
+    Fire: 0
+    Acid: 0
+    Magic: 0
+    Bio: 0
 --- !u!1 &4827224533573323059
 GameObject:
   m_ObjectHideFlags: 0
@@ -246,7 +269,7 @@ GameObject:
   - component: {fileID: 3493423571535560689}
   - component: {fileID: 3651291330810827203}
   - component: {fileID: 3708589072684735304}
-  - component: {fileID: 8310532552690989649}
+  - component: {fileID: 8135233970803992685}
   m_Layer: 12
   m_Name: NPC_Xenomorph
   m_TagString: Untagged
@@ -452,6 +475,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  isDirty: 0
   sceneId: 0
   serverOnly: 0
   m_AssetId: 
@@ -556,11 +580,12 @@ MonoBehaviour:
   followTarget: {fileID: 0}
   spriteHolder: {fileID: 3392343644113499114}
   targetOtherPlayersWhoGetInWay: 1
+  onlyHitTarget: 0
   hitDamage: 15
   attackVerb: slashed
   defaultTarget: 0
   meleeCoolDown: 0.4
---- !u!114 &8310532552690989649
+--- !u!114 &8135233970803992685
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -569,19 +594,16 @@ MonoBehaviour:
   m_GameObject: {fileID: 4827737755655954553}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ba86e59112fcacd448eb7b1203dd37f1, type: 3}
+  m_Script: {fileID: 11500000, guid: 243a3c078e195af4dadbc9f673627997, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   mobName: Xenomorph
   knockedDownRotation: -90
-  searchTickRate: 0.5
-  DeathSounds:
+  deathSounds:
   - xenodie
-  GenericSounds:
-  - Hiss1
-  - Hiss2
-  - Hiss3
-  - Hiss4
-  - Hiss5
-  PlaySoundTime: 3
-  status: 0
+  randomSound:
+  - Hiss#
+  playRandomSoundTimer: 3
+  randomSoundProbability: 20
+  searchTickRate: 0.5
+  currentStatus: 0
