+++ UnityProject/Assets/Resources/Prefabs/SceneConstruction/NestedManagers/SoundManager.prefab
@@ -4443,7 +4443,7 @@ GameObject:
   - component: {fileID: 4794713589522012633}
   - component: {fileID: 3043858601886721449}
   m_Layer: 0
-  m_Name: god1
+  m_Name: God1
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -14398,7 +14398,7 @@ GameObject:
   - component: {fileID: 7551001907513119286}
   - component: {fileID: 2582926923459114014}
   m_Layer: 0
-  m_Name: ghost
+  m_Name: Ghost
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -21061,7 +21061,7 @@ GameObject:
   - component: {fileID: 2658236004763317763}
   - component: {fileID: 3444732619845469160}
   m_Layer: 0
-  m_Name: xenodie
+  m_Name: Xenodie
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -29809,7 +29809,7 @@ GameObject:
   - component: {fileID: 7178878014952527575}
   - component: {fileID: 4195209845306206608}
   m_Layer: 0
-  m_Name: statuedeath
+  m_Name: StatueDeath
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -33779,7 +33779,7 @@ GameObject:
   - component: {fileID: 6312314817849658030}
   - component: {fileID: 5412501998241263617}
   m_Layer: 0
-  m_Name: facehugger
+  m_Name: Facehugger
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -33943,7 +33943,7 @@ GameObject:
   - component: {fileID: 1873993681756744023}
   - component: {fileID: 7324835528026017541}
   m_Layer: 0
-  m_Name: god2
+  m_Name: God2
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -36968,7 +36968,7 @@ GameObject:
   - component: {fileID: 8006151980984199213}
   - component: {fileID: 6274634402762547776}
   m_Layer: 0
-  m_Name: dragon
+  m_Name: Dragon
   m_TagString: SoundFX
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
