--- UnityProject/Assets/Scripts/NPC/AI/ClownAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f423e516d0a62384493c3095abe85efa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
