+++ UnityProject/Assets/Scripts/NPC/AI/Friendly/CorgiAI.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b9b02e30a4b623144905790b2eaf49b9
+guid: 715d7edbfa5ae6742b05a85634ad752c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
