--- UnityProject/Assets/Scripts/NPC/AI/GodAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a9e36df76b32ffd489ecff328e122570
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
