+++ UnityProject/Assets/Scripts/NPC/AI/Hostile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 62c5b1e8133d453580b4b5ecea60788d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
