+++ UnityProject/Assets/Scripts/NPC/AI/Hostile/GenericHostileAI.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba86e59112fcacd448eb7b1203dd37f1
+guid: 243a3c078e195af4dadbc9f673627997
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
