+++ UnityProject/Assets/Scripts/NPC/AI/Hostile/StatueAI.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5cdd796a9bd8bb04c8a114e9371f0e89
+guid: 526f944188797b340b690113f0b7db00
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
